backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
sentence-transformers==2.7.0
pinecone[grpc]>=8.0.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0